Stop Listener from double watching directories.
Listener will watch all files in a directory recursively by default for changes to automate updates to webservers. We had two symlinks in our project that caused errors in the logs from Listener seeing those directories twice. But the symlinks were kind of strange. Because of the dependency on the exact code in www.getdbt.com, I renamed images img and updated references. The result is a proper use of these models, clean logs, and removal of symlinks.